% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{del_sigmoid}
\alias{del_sigmoid}
\title{Derivative of sigmoid function}
\usage{
del_sigmoid(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of solutions of \code{sigmoid(x)*(1-sigmoid(x))}
}
\description{
Calculating the sigmoid's first derivative of a vector
}
\examples{
del_sigmoid(1)
del_sigmoid(1:100)
}
