% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1.R
\name{del_tanh}
\alias{del_tanh}
\title{Derivative of tangens hyperbolicus}
\usage{
del_tanh(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of solutions of \code{4/((exp(-x)+exp(x))^2)}
}
\description{
Calculating the tangens hyperbolicus' first derivative of a vector
}
\examples{
del_tanh(1)
del_tanh(1:100)
}
