% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{load_NN}
\alias{load_NN}
\title{Load the neural network}
\usage{
load_NN(path = "R/sysdata.rda")
}
\arguments{
\item{path}{path from working directory to the saved file which is in NNPackage/R/dateiname}
}
\description{
Loading in the saved neural network including it's weights
}
