% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1.R
\name{NN}
\alias{NN}
\title{S6 class that can generate Neural Networks}
\description{
S6 class that can generate Neural Networks

S6 class that can generate Neural Networks
}
\details{
In this class Neural Networks can be generated and optimized via different
Gradient Descends Methods

\describe{ \item{L}{number of hidden layers}

\item{B}{width of each layer}

\item{W}{weight matrices}

\item{d}{width vector} }
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-f}{\code{NN$f()}}
\item \href{#method-del_f}{\code{NN$del_f()}}
\item \href{#method-new}{\code{NN$new()}}
\item \href{#method-ffprop}{\code{NN$ffprop()}}
\item \href{#method-ffprop_clas}{\code{NN$ffprop_clas()}}
\item \href{#method-eval_till_layer}{\code{NN$eval_till_layer()}}
\item \href{#method-eval_till_layer_z}{\code{NN$eval_till_layer_z()}}
\item \href{#method-GD3}{\code{NN$GD3()}}
\item \href{#method-BP_reg}{\code{NN$BP_reg()}}
\item \href{#method-SGD}{\code{NN$SGD()}}
\item \href{#method-SGD_clas}{\code{NN$SGD_clas()}}
\item \href{#method-BP_clas}{\code{NN$BP_clas()}}
\item \href{#method-clone}{\code{NN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-f"></a>}}
\if{latex}{\out{\hypertarget{method-f}{}}}
\subsection{Method \code{f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$f(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-del_f"></a>}}
\if{latex}{\out{\hypertarget{method-del_f}{}}}
\subsection{Method \code{del_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$del_f(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializing a NN
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$new(
  L = 1,
  B = c(1, 1, 1),
  W = c(1, 1, 1),
  d = c(1, 1, 0),
  min_gewicht = -2,
  max_gewicht = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{L}}{a vector}

\item{\code{B}}{a vector}

\item{\code{W}}{a vector}

\item{\code{d}}{a vector}

\item{\code{min_gewicht}}{a vector}

\item{\code{max_gewicht}}{a vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of solutions of \code{x}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ffprop"></a>}}
\if{latex}{\out{\hypertarget{method-ffprop}{}}}
\subsection{Method \code{ffprop()}}{
Calculating the function of NN
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$ffprop(x = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ffprop_clas"></a>}}
\if{latex}{\out{\hypertarget{method-ffprop_clas}{}}}
\subsection{Method \code{ffprop_clas()}}{
Calculating feed forward with softmax
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$ffprop_clas(x = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_till_layer"></a>}}
\if{latex}{\out{\hypertarget{method-eval_till_layer}{}}}
\subsection{Method \code{eval_till_layer()}}{
Calculating NN up to a certain layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$eval_till_layer(x = 1, Layer = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector}

\item{\code{L}}{A vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_till_layer_z"></a>}}
\if{latex}{\out{\hypertarget{method-eval_till_layer_z}{}}}
\subsection{Method \code{eval_till_layer_z()}}{
Calculating NN up to a certain layer without activation function in last step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$eval_till_layer_z(x = 1, Layer = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector}

\item{\code{L}}{A vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GD3"></a>}}
\if{latex}{\out{\hypertarget{method-GD3}{}}}
\subsection{Method \code{GD3()}}{
Calculating gradient descent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$GD3(x, y, iteration = 10, delta = 0.02)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector}

\item{\code{y}}{A vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BP_reg"></a>}}
\if{latex}{\out{\hypertarget{method-BP_reg}{}}}
\subsection{Method \code{BP_reg()}}{
Calculating gradient descent for regression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$BP_reg(x, y, gam = 1e-04, lambda = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector}

\item{\code{y}}{A vector}

\item{\code{gam}}{A vector}

\item{\code{lambda}}{A vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SGD"></a>}}
\if{latex}{\out{\hypertarget{method-SGD}{}}}
\subsection{Method \code{SGD()}}{
Calculating stochastic gradient descent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$SGD(x, y, n, delta = 0.01, iteration = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector}

\item{\code{y}}{A vector}

\item{\code{n}}{a vector}

\item{\code{delta}}{a vector}

\item{\code{iteration}}{a vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SGD_clas"></a>}}
\if{latex}{\out{\hypertarget{method-SGD_clas}{}}}
\subsection{Method \code{SGD_clas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$SGD_clas(x, y, n, delta = 0.01, iteration = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BP_clas"></a>}}
\if{latex}{\out{\hypertarget{method-BP_clas}{}}}
\subsection{Method \code{BP_clas()}}{
Calculating gradient descent for classification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$BP_clas(x, y, gam = 1e-04, lambda = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector}

\item{\code{y}}{A vector}

\item{\code{gam}}{A vector}

\item{\code{lambda}}{A vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
