% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNPackage-package.R
\docType{package}
\name{NNPackage-package}
\alias{NNPackage}
\alias{NNPackage-package}
\title{NNPackage: A neural network package that can generate and optimize neural networks}
\description{
With this package it is possible to generate neural networks with different layer depth and different neurons per layer.
    Additionally the pacakge includes 3 different optimizers so that the neural network can be trained for a given data set.
}
\author{
\strong{Maintainer}: Mark Daniel Arndt \email{markdaniel.arndt@gmail.com}

}
\keyword{internal}
