% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{save_NN}
\alias{save_NN}
\title{Save the neural network}
\usage{
save_NN(NN)
}
\arguments{
\item{NN}{A NN-Class Object}
}
\description{
Saving the neural network including it's weights
}
\examples{
save_NN(NN$new(B=c(1,10,1)))
}
