% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1.R
\name{sigmoid}
\alias{sigmoid}
\title{Calculating the sigmoid of a vector}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of solutions of \code{1/(1+exp(-x))}
}
\description{
Calculating the sigmoid of a vector
}
\examples{
sigmoid(1)
sigmoid(1:100)
}
