% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1.R
\name{softmax}
\alias{softmax}
\title{Softmax function}
\usage{
softmax(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of solutions of \code{1/sum(exp(x))*exp(x)}
}
\description{
Calculating the softmax function of a vector
}
\details{
The softmax function maps the entries a vector onto the intervall (0,1]
}
\examples{
softmax(1)
softmax(1:100)
}
