% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tanh}
\alias{tanh}
\title{Tangens hyperbolicus}
\usage{
tanh(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of solutions of \code{(exp(2*x)-1)/(exp(2*x)+1)}
}
\description{
Calculating the tangens hyperbolicus of a vector
}
\examples{
tanh(1)
tanh(1:100)
}
