% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1.R
\name{NN}
\alias{NN}
\title{The title for my S4 class that extends \code{"character"} class.}
\description{
Some details about this class and my plans for it in the body.
}
\details{
\describe{
\item{myslot1}{A logical keeping track of something.}

\item{myslot2}{An integer specifying something else.}

\item{myslot3}{A data.frame holding some data.}
}
}
\examples{

## ------------------------------------------------
## Method `NN$calculate2`
## ------------------------------------------------

id(1)
id(1:100)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-f}{\code{NN$f()}}
\item \href{#method-del_f}{\code{NN$del_f()}}
\item \href{#method-new}{\code{NN$new()}}
\item \href{#method-calculate2}{\code{NN$calculate2()}}
\item \href{#method-cal_clas}{\code{NN$cal_clas()}}
\item \href{#method-eval_till_layer}{\code{NN$eval_till_layer()}}
\item \href{#method-eval_till_layer_z}{\code{NN$eval_till_layer_z()}}
\item \href{#method-BP_reg}{\code{NN$BP_reg()}}
\item \href{#method-GD}{\code{NN$GD()}}
\item \href{#method-GD2}{\code{NN$GD2()}}
\item \href{#method-GD3}{\code{NN$GD3()}}
\item \href{#method-SGD}{\code{NN$SGD()}}
\item \href{#method-GD_clas}{\code{NN$GD_clas()}}
\item \href{#method-BP_klas}{\code{NN$BP_klas()}}
\item \href{#method-clone}{\code{NN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-f"></a>}}
\if{latex}{\out{\hypertarget{method-f}{}}}
\subsection{Method \code{f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$f(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-del_f"></a>}}
\if{latex}{\out{\hypertarget{method-del_f}{}}}
\subsection{Method \code{del_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$del_f(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$new(
  L = 1,
  B = c(1, 1, 1),
  W = c(1, 1, 1),
  d = c(1, 1, 0),
  min_gewicht = -2,
  max_gewicht = 2
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate2"></a>}}
\if{latex}{\out{\hypertarget{method-calculate2}{}}}
\subsection{Method \code{calculate2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$calculate2(x = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of solutions of \code{x}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{id(1)
id(1:100)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_clas"></a>}}
\if{latex}{\out{\hypertarget{method-cal_clas}{}}}
\subsection{Method \code{cal_clas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$cal_clas(x = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_till_layer"></a>}}
\if{latex}{\out{\hypertarget{method-eval_till_layer}{}}}
\subsection{Method \code{eval_till_layer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$eval_till_layer(x = 1, Layer = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_till_layer_z"></a>}}
\if{latex}{\out{\hypertarget{method-eval_till_layer_z}{}}}
\subsection{Method \code{eval_till_layer_z()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$eval_till_layer_z(x = 1, Layer = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BP_reg"></a>}}
\if{latex}{\out{\hypertarget{method-BP_reg}{}}}
\subsection{Method \code{BP_reg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$BP_reg(x, y, gam = 1e-04, lambda = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GD"></a>}}
\if{latex}{\out{\hypertarget{method-GD}{}}}
\subsection{Method \code{GD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$GD(x, y, lambda = 1, stepsize = 1e-04, iterations = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GD2"></a>}}
\if{latex}{\out{\hypertarget{method-GD2}{}}}
\subsection{Method \code{GD2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$GD2(x, y, lambda = 1, stepsize = 1e-04, iterations = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GD3"></a>}}
\if{latex}{\out{\hypertarget{method-GD3}{}}}
\subsection{Method \code{GD3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$GD3(x, y, iteration = 10, delta = 0.02)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SGD"></a>}}
\if{latex}{\out{\hypertarget{method-SGD}{}}}
\subsection{Method \code{SGD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$SGD(x, y, n, delta = 0.01, iteration = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GD_clas"></a>}}
\if{latex}{\out{\hypertarget{method-GD_clas}{}}}
\subsection{Method \code{GD_clas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$GD_clas(x, y, iteration = 1000, delta = 0.002)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BP_klas"></a>}}
\if{latex}{\out{\hypertarget{method-BP_klas}{}}}
\subsection{Method \code{BP_klas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$BP_klas(x, y, gam = 1e-04, lambda = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
